Sync phased array branch to trunk

+++ sfxc/utils/run_jobs.py
@@ -25,18 +25,20 @@ def time2vex(secs):
     tupletime = time.gmtime(secs)
     return time.strftime("%Yy%jd%Hh%Mm%Ss", tupletime)
 
-usage = "usage: %prog [options] vexfile"
+usage = "usage: %prog [options] vexfile controlfiles"
 parser = optparse.OptionParser(usage=usage)
 parser.add_option("-n", "--nodes", dest="number_nodes",
                   default=256, type="int",
                   help="Number of correlator nodes",
                   metavar="N")
+parser.add_option("-d", "--no-delays", dest="gen_delays",
+                  default = True,
+                  action = "store_false",
+                  help="Disable generating new delays")
 parser.add_option("-m", "--machines", dest="machines",
                   default="a,b,c,d,e,f,g,h,i,j", type="string",
                   help="Machines to run correlator nodes on",
                   metavar="LIST")
-parser.add_option('-f', '--file', dest='files', type="string",
-                  help = "Stations that are passed as a file, usage : STATION1:SFXC_NODE,STATION2:SFXC_NODE,...")
 (options, args) = parser.parse_args()
 
 if len(args) < 2:
@@ -44,14 +46,6 @@ if len(args) < 2:
     pass
 
 vex_file = args[0]
-
-
-files = {}
-if options.files:
-  f = options.files.split(',')
-  for val in f:
-    station,node = val.split(':')
-    files[station] = 'sfxc-'+node+'.sfxc'
 # Parse the VEX file.
 vex = Vex(vex_file)
 exper = vex['GLOBAL']['EXPER']
@@ -59,34 +53,8 @@ exper = vex['GLOBAL']['EXPER']
 # Proper time.
 os.environ['TZ'] = "UTC"
 
-##mk5s = ['mk5-' + str(x) for x in range(17)]
-#mk5s = ['10.88.1.' + str(200+x) for x in range(17)]
-print 'Warning : overidden Mark5s'
-mk5s = [
-#'10.88.1.200', #mk5-0
-#'10.88.1.201', #mk5-1
-#'10.88.1.202', #mk5-2
-#'10.88.1.203', #mk5-3
-#'10.88.1.204', #mk5-4
-#'10.88.1.205', #mk5-5
-'10.88.1.206', #mk5-6
-#'10.88.1.207',  #mk5-7
-#'10.88.1.220', # mk5-c0
-#'10.88.1.221', # mk5-c1
-#'10.88.1.222', # mk5-c2
-#'10.88.1.223', # mk5-c3
-#'10.88.1.224', # mk5-c4
-#'10.88.1.225', # mk5-c5
-#'10.88.1.208', #mk5-8
-#'10.88.1.209', #mk5-9
-'10.88.1.210', #mk5-10
-'10.88.1.211', #mk5-11
-'10.88.1.212', #mk5-12
-#'10.88.1.213', #mk5-13
-'10.88.1.214', #mk5-14
-'10.88.1.215', #mk5-15
-'10.88.1.216'] #mk5-16
-
+mk5s = ['10.88.1.' + str(200+x) for x in range(17)]
+mk5s += ['10.88.1.' + str(200+x) for x in range(20,26)]
 manager_node = "head.sfxc"
 output_node = "head.sfxc"
 log_node = "head.sfxc"
@@ -116,7 +84,6 @@ for machine in options.machines.split(','):
     continue
 
 for ctrl_file in args[1:]:
-
     basename = os.path.splitext(os.path.basename(ctrl_file))[0]
     machine_file = basename + ".machines"
     rank_file = basename + ".ranks"
@@ -125,7 +92,8 @@ for ctrl_file in args[1:]:
     fp = open(ctrl_file, 'r')
     json_input = json.load(fp)
     fp.close()
-
+    if 'file_parameters' not in json_input:
+      json_input['file_parameters'] = {}
     start = vex2time(json_input['start'])
 
     stations = json_input['stations']
@@ -133,39 +101,40 @@ for ctrl_file in args[1:]:
 
     # Make sure the delay files are up to date, and generate new ones
     # if they're not.
-    procs = {}
-    success = True
-    for station in stations:
-        path = urlparse.urlparse(json_input['delay_directory']).path
-        delay_file = path + '/' +  exper + '_' + station + '.del'
-        if not os.access(delay_file, os.R_OK) or \
-           os.stat(delay_file).st_mtime < os.stat(vex_file).st_mtime:
-            args = ['generate_delay_model', vex_file, station, delay_file]
-            procs[station] = subprocess.Popen(args, stdout=subprocess.PIPE)
-            pass
-        continue
-    for station in procs:
-        output = procs[station].communicate()[0]
-        procs[station].wait()
-        if procs[station].returncode != 0:
-            print "Delay model couldn't be generated for " + station + ":"
-            print output
-            path = urlparse.urlparse(json_input['delay_directory']).path
-            delay_file = path + '/' +  exper + '_' + station + '.del'
-            os.remove(delay_file)
-            success = False
-            pass
-        continue
-    if not success:
-        sys.exit(1)
-        pass
+    if options.gen_delays:
+      procs = {}
+      success = True
+      for station in stations:
+          path = urlparse.urlparse(json_input['delay_directory']).path
+          delay_file = path + '/' +  exper + '_' + station + '.del'
+          if not os.access(delay_file, os.R_OK) or \
+             os.stat(delay_file).st_mtime < os.stat(vex_file).st_mtime:
+              args = ['generate_delay_model', vex_file, station, delay_file]
+              procs[station] = subprocess.Popen(args, stdout=subprocess.PIPE)
+              pass
+          continue
+      for station in procs:
+          output = procs[station].communicate()[0]
+          procs[station].wait()
+          if procs[station].returncode != 0:
+              print "Delay model couldn't be generated for " + station + ":"
+              print output
+              path = urlparse.urlparse(json_input['delay_directory']).path
+              delay_file = path + '/' +  exper + '_' + station + '.del'
+              os.remove(delay_file)
+              success = False
+              pass
+          continue
+      if not success:
+          sys.exit(1)
+          pass
 
     # Figure out the directory where the input data files are located.
     # Should be the same for all stations otherwise things will become
     # way too complicated.
     data_dir = None
     for station in stations:
-      if station not in files:
+      if station not in json_input['file_parameters']:
         data_source = json_input['data_sources'][station][0]
         if urlparse.urlparse(data_source).scheme == 'file':
             path = urlparse.urlparse(data_source).path
@@ -240,7 +209,7 @@ for ctrl_file in args[1:]:
         # chance to mount the missing media.
         missing = False
         for station in stations:
-          if station not in files:
+          if station not in json_input['file_parameters']:
             if not station in input_nodes:
                 if not missing:
                     print "Please mount media with the following VSNs:"
@@ -259,22 +228,36 @@ for ctrl_file in args[1:]:
             pass
         continue
 
+    # Initialize ranks
+    ranks = {}
     # Create a MPI machine file for the job.
     fp = open(machine_file, 'w')
     print >>fp, manager_node, "slots=4" #Assume manager, output, and log node on the same machine
     #print >>fp, output_node
     #print >>fp, log_node
     for station in stations:
-      if station not in files:
+      if station not in json_input['file_parameters']:
         #ifhn = "ifhn="+input_nodes[station]
 	print >>fp, " #", station
         print >>fp, input_nodes[station], " slots=4"
-        continue
+        ranks[input_nodes[station]]  = 4
+    for station in json_input['file_parameters']:
+      source = json_input['file_parameters'][station]['sources'][0]
+      machine = source.partition(':')[0]
+      if machine.startswith('sfxc-'):
+        if not machine.endswith('.sfxc'):
+          machine += '.sfxc'
+          if machine not in machines:
+            print >>fp, machine, " slots=8"
+            ranks[machine] = 8
+      elif machine.startswith('aribox') or (machine == '10.88.1.230'):
+        print >>fp, machine, " slots=16"
+        ranks[machine] = 16
+
     #for i in range(8):
     for machine in machines:
         print >>fp, machine, " slots=8"
-        continue
-    #    continue
+        ranks[machine] = 8
     fp.close()
 
     # Create a MPI rank file for the job.
@@ -284,23 +267,25 @@ for ctrl_file in args[1:]:
     print >>fp, "rank 2=", log_node, "slot=2,3"
     rank=2
     # Create ranks
-    ranks = {}
-    for machine in machines:
-      ranks[machine] = 8
     for station in stations:
       rank += 1
-      if station in files:
-        node = files[station]
-        print >>fp, "rank", str(rank), "=", node, "slot=4,5,6,7"
-        ranks[node] -= 4
+      if station in json_input['file_parameters']:
+        source = json_input['file_parameters'][station]['sources'][0]
+        node = source.partition(':')[0]
+        if node.startswith('sfxc-') and not node.endswith('.sfxc'):
+            node += '.sfxc'
+        nthread = json_input['file_parameters'][station]['nthreads']
+        slots = 'slot=' + ','.join([str(i) for i in range(ranks[node]-nthread, ranks[node])])
+        ranks[node] -= nthread
+        print >>fp, "rank", str(rank), "=", node, slots
       else:
-        print >>fp, "rank", str(rank), "=", input_nodes[station], "slot=0,2"
+        print >>fp, "rank", str(rank), "=", input_nodes[station], "slot=0,1,2"
 
     for i in range(8):
         for machine in machines:
           if ranks[machine] > 0:
             rank += 1
-            print >>fp, "rank", str(rank), "=", machine, "slot=", str(i)
+            print >>fp, "rank", str(rank), "=", machine, "slot=", str(ranks[machine]-1)
             ranks[machine] -= 1
     fp.close()
